Intermittent Save
This is NOT an official publishing of these notes/blog posts. Instead,
it is an intermittent save so as to safeguard what has already been put
down. I am trying to avoid what happened to the original Official Blog
Post #4 by doing this. These will all be resaved and edited soon.